refactor(waveguide): extract pure-NumPy mode solver to _waveguide_modes.py
waveguide_port.py was a 3364-LOC multi-concern module. Extract the
self-contained, pure-NumPy transverse mode solver + mode-profile algebra
(10 functions, lines 359-648: _second_diff_1d, _galerkin_*,
_cell_centred_gradient, _pick_eigenmode_*, _aperture_area,
_shift_profile_to_dual, _orthonormalize_profile_arrays,
_scale_h_to_unit_cross) into a sibling module rfx/sources/_waveguide_modes.py
and re-import them back into the waveguide_port namespace.

The block is off the FDTD/JAX AD tape (all float64 NumPy, runs at port
setup) and references only numpy — no jnp/EPS_0/MU_0/C0_LOCAL/rfx coupling,
so the new module imports only numpy and cannot introduce an import cycle.
This is a verbatim move (behavior-preserving), NOT the speculative 7-module
package split (refuted as churn: _compute_beta/_plane_*/C0_LOCAL stitch the
extraction clusters together and relocating them risks the AD tape).

waveguide_port.py 3364 -> 3090 LOC; new module 311 LOC.

Bit-identity gate (all green):
- Mode-solver byte-identity: SHA256 of the 10 functions' outputs on fixed
  inputs is identical pre/post (b825bc54...dad518).
- Import-compat: all public + private re-exports resolve from
  rfx.sources.waveguide_port unchanged, incl. eigenmode.py's private deps
  (_te_mode_profiles/_tm_mode_profiles/cutoff_frequency/C0_LOCAL, which were
  NOT moved); the moved helpers are the SAME object via both paths.
- 57 waveguide/eigenmode/multimode/AD tests pass (AD tape unaffected).
- New tests/test_waveguide_modes_extraction_contract.py pins the re-export
  surface + the pure-NumPy invariant against future drift.
